Fix bugs in `memcmp` and refactoring code
