+++ bootloader/src/main.c
@@ -31,19 +31,11 @@
 
 void main()
 {
-    // set up serial console
-    // uart_init();
     shell_init();
     while(1) {
         uart_puts("# ");
         char cmd[CMD_LEN];
         shell_input(cmd);
         shell_controller(cmd);
-
-        // c=uart_getc();
-        // uart_send(c);
-        // uart_puts("\n\n");
-        // if(c=='1') power_off();
-        // if(c=='2') reset();
     }
 }
