+++ include/file.h
@@ -1,2 +1,1 @@
 void readfile(char *target, int file_size);
-void readline(char *target, char *buffer, unsigned long buff_size);
