+++ include/initrd.h
@@ -22,7 +22,24 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-extern volatile unsigned char _binary_ramdisk_start;
+// extern volatile unsigned char _binary_ramdisk_start;
 void initrd_list(char *buf);
 void initrd_ls(char *buf);
 void initrd_cat(char *buf);
+
+typedef struct {
+    char magic[6];
+    char ino[8];
+    char mode[8];
+    char uid[8];
+    char gid[8];
+    char nlink[8];
+    char mtime[8];
+    char filesize[8];
+    char devmajor[8];
+    char devminor[8];
+    char rdevmajor[8];
+    char rdevminor[8];
+    char namesize[8];
+    char check[8];
+} __attribute__((packed)) cpio_f;
