+++ src/file.c
@@ -1,12 +1,7 @@
 #include "uart.h"
 
-
 void readfile(char *target, unsigned long file_size)
 {
     while (file_size-- > 0)
         uart_send(*(target++));
-}
-void readline(char *target, char *buffer, unsigned long buff_size)
-{
-
 }
