+++ src/my_stdlib.c
@@ -1,0 +1,9 @@
+
+
+
+int memcmp(void *s1, void *s2, int n)
+{
+    unsigned char *a=s1,*b=s2;
+    while(n-->0){ if(*a!=*b) { return *a-*b; } a++; b++; }
+    return 0;
+}
