+++ src/shell.c
@@ -7,15 +7,13 @@
 void shell_init()
 {
     uart_init();
-    uart_puts("\n\n Hello from Raspi 3\n");
+    uart_puts("\n\n Hello from Raspi 3b+\n");
 }
 
 void shell_input(char *cmd)
 {
     char c;
     int idx = 0, end = 0;
-
-    // uart_puts("\r# ");
 
     while ((c = uart_getc()) != '\n') {
         uart_send(c);
@@ -35,7 +33,6 @@ void shell_controller(char *cmd)
         uart_puts("hello     : print Hello World!\n");
         uart_puts("ls        : list file in initramfs.cpio\n");
         uart_puts("cat       : show the file content\n");
-        // uart_puts("timestamp: get current timestamp\n");
         uart_puts("reboot    : reboot the device\n");
     } else if (!strcmp(cmd, "hello")) {
         uart_puts("Hello World!\n");
